requirements.txt
streamlit
python-dotenv
crewai
crewai_tools
pandas
plotly
matplotlib
seaborn
fpdf
langchain_groq
langchain
langchain-community
faiss-cpu
sentence-transformers